Fix: change requires_role_functions

+++ app/routers/company.py
@@ -22,18 +22,17 @@ def find_all_company_with_filter(*, search_kw: str = Query(default=""),
                                         page_number=page_number,
                                         page_size=page_size,
                                         )
-    
-@requires_system_admin    
 @router.post("", status_code=status.HTTP_201_CREATED, response_model=CompanyViewModel)
 def create_company(request: CompanyPostModel,
                    db: Session = Depends(get_db_context),
                    user: User = Depends(AuthService.token_interceptor)):
+    requires_system_admin(user)
     return CompanyService.create_company(db, request)
 
-@requires_system_admin
 @router.put("/{id}", response_model=CompanyViewModel)
 def update_company(id: UUID, request: CompanyPostModel, db: Session = Depends(get_db_context), 
                    user: User = Depends(AuthService.token_interceptor)):
+    requires_system_admin(user)
     if id == UUID(SYSTEM_COMPANY_ID) or id == UUID(NONE_COMPANY_ID):
         raise ForbiddenOperationException
     return CompanyService.update_company(db, id, request)
@@ -45,10 +44,10 @@ def get_one_company(id: UUID, db: Session = Depends(get_db_context)):
         raise ResourceNotFoundException("Company")
     return company
 
-@requires_system_admin
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT, response_model=None)
 def delete_company(id: UUID, db: Session = Depends(get_db_context), 
                    user: User = Depends(AuthService.token_interceptor)):
+    requires_system_admin(user)
     if id == UUID(SYSTEM_COMPANY_ID) or id == UUID(NONE_COMPANY_ID):
-        raise ForbiddenOperationException
+        raise ForbiddenOperationException()
     CompanyService.delete_company(db, id)
